Refactor App component and import dependencies
