+++ devhub/projects/udemy_react_the_complete_guide_2024/s03/src/App.jsx
@@ -1,6 +1,6 @@
 import componetsImg from "./assets/components.png";
 import { CORE_CONCEPTS } from "./data.js";
-import Header from "./components/Header.jsx";
+import Header from "./components/Header/Header.jsx";
 import CoreConcept from "./components/CoreConcept.jsx";
 
 function App() {
