+++ devhub/projects/udemy_react_the_complete_guide_2024/s03/src/components/Header/Header.jsx
@@ -1,4 +1,5 @@
-import reactImg from "../assets/react-core-concepts.png";
+import reactImg from "../../assets/react-core-concepts.png";
+import "./Header.css";
 
 const reactDescription = ["Fundamental", "Crucial", "Core"];
 function genRandomInt(max) {
